Added probabilities word method
Added methods to calculate probabilities of the lenght base word and
probability that additional word will be a alphabetical.

Since we have a name length of 12 characters, an additional word with
the desired length may not exist in the dictionary; in this case, an
additional word will be a set of numbers.

* Added `get_probabilities_len_base_word` method
* Added `get_probability_alphabet_additional_word` method