Implementation `generate` method in RandomNameGenerator
Added a method for generating a name based on the base word, the
probability of the length of the base word depends on their frequency.

An additional word can consist of letters or numbers, the probability of
getting numbers sets `numbers_probabilities` variable and existing
additional alphabetical word in base dictionary.

* Added implementation of the `generate` method
* Added implementation of the `generate_list` method
* Added `get_random_name` method
* Update test cases