Implement `random_provider` to `RandomNameGenerator`
Added the ability to integrate your own `random_provider` which
will be faster than the build-in (for example `random`, or `numpy`).

Two providers are implemented for selection:
* numpy.random
* FastRandomChoice

`FastRandomChoice` - is a class extend build-in `random` python
package. Its provide `choice` method with probabilities selection.

`FastRandomChoice.choice()` is 5 times faster than `numpy.random.choice`.
Data was checked on a `5_000_000` sample.